+++ governor/src/quota.rs
@@ -1,8 +1,6 @@
-use std::prelude::v1::*;
-
+use core::num::NonZeroU32;
+use core::time::Duration;
 use nonzero_ext::nonzero;
-use std::num::NonZeroU32;
-use std::time::Duration;
 
 use crate::nanos::Nanos;