+++ governor/src/state/keyed/dashmap.rs
@@ -1,13 +1,11 @@
 #![cfg(all(feature = "std", feature = "dashmap"))]
 
-use std::prelude::v1::*;
-
 use crate::nanos::Nanos;
 use crate::state::{InMemoryState, StateStore};
 use crate::{clock, Quota, RateLimiter};
 use crate::{middleware::NoOpMiddleware, state::keyed::ShrinkableKeyedStateStore};
+use core::hash::Hash;
 use dashmap::DashMap;
-use std::hash::Hash;
 
 /// A concurrent, thread-safe and fairly performant hashmap based on [`DashMap`].
 pub type DashMapStateStore<K> = DashMap<K, InMemoryState>;